[MISC] Agregado de imágenes default y textos default de prueba para testeo en categoría/imágenes

+++ src/ar/uba/fi/talker/dao/ImageTalkerDataSource.java
@@ -83,7 +83,7 @@ public class ImageTalkerDataSource {
 		return image;
 	}
 	
-	public List<ImageDAO> getImagesForCategory(int keyId) {
+	public List<ImageDAO> getImagesForCategory(long keyId) {
 		List<ImageDAO> images = new ArrayList<ImageDAO>();
 		
 		Cursor cursor = database.rawQuery("SELECT * FROM "
